src/requirements.txt
h5py<3.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
matplotlib==3.3.2
mtcnn==0.1.0
numpy==1.18.5
opencv-python==4.4.0.42
pandas==1.1.2
pylint==2.6.0
scipy==1.4.1
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
pytest~=6.2.2
joblib~=1.0.1
scikit-learn~=0.24.1
Pillow~=8.1.1
boto3==1.17.68
